Fix: Resolve Map Markers Disabling On Update
There was an issue where if you had previously chosen to enable
map markers, and then updated to a newer version of Skyshards,
it would disable the map markers. This fix prevents that by
checking for any previously enabled map markers before the update
and then persisting that state to a global variable.