Fix: Configurator Options Don't Always Work
There were instances where a player could be in possession of an
outdated configurator which doesn't have updated properties which
would cause certain options to fail when selected.

This fix tracks the version of the configurator going forward and
links it to the Skyshards script version running. Any time it
detects a version mismatch it will update the configurator.

Additionally, older configurator versions will be detected and
removed from the player's inventory to prevent such issues.

This fix will also be applied retroactively to saves below v1.6.0
once per game load, updating the old configurator.

Finally, an extra configurator option was added to reset the
configurator should a player have the need to manually do so. It
will remove the current configurator and give the player a new one.