Feat: Add Ability To Persist God Mode
Previously god mode would be disabled after the absorb animation
ended, which could lead to a negative player experience. Now the
player has an option to persist god mode in the configurator which
will not touch the state of god mode if enabled.

The reason god mode was enabled/disabled to begin with was to ensure
that the player was not attacked and possibly killed during the
Skyshard absorption animation.