Fix: Disable Player Controls Without Conflict
Prior to this update all player controls would be disabled as part
of the Skyshard absorption animation and then possibly incorrectly
re-enabled if a player had chosen to disable certain ones prior via
other means.

This update only disables two specific controls; combat and
movement, and leaves the rest of the controls untouched. This should
result in far greater compatibility with user and mod settings.