Fix: Move Skyshard Conflicting With Snowbound Acres Mod
A specific Skyshard in Falkreath was conflicting with another mod,
Snowbound Acres, as the author made landscape changes which left
the poor Skyshard floating mid-air! This fix firmly plants the
Skyshard in a new home out of harms way.

Hopefully mod authors will refrain from making landscape and/or
world changes to the base game and rather integrate their mods
into the world instead of expecting the world to integrate into
their mod. This would allow for better compatibility with a greater
variety of mods.