Fix: Absorption Function Reworked To Avoid Perk Point Lockout
Previously the function would only check if the player had the same
amount of absorptions as the absorption cap, and if it was met
the player would receive the perk points as specified.

However, there are certain cases where the absorption cap
could change to be less than the current absorption value and
thus the player ends up in a situation where they no longer receive
perk points going forward.

This change not only handles the checking better, it also has a
loop included to hand out perk points so long as the current
value is greater than or equal to the absorption cap.

Furthermore, checks were added to prevent any overflow issues that
could arise under certain use cases.