Refactor: Update Side Quest Progress Synchronously
Previously side quest progression happened asynchronously which
meant that it could take a while before the specific Skyshard
quest was updated. This massive update introduces helper threads for
each specific side quest that can all be run at the same time and
update their respective quests in a non-blocking manner.

A quest manager was also introduced to both simplify and speed up
the process and a primary quest helper added that ensures the
correct flow of quest updating as well as performing any other
necessary checks and updates.

This change is backwards compatible with the previous system.