Refactor: Better Handle Stopping Of Quest Tracking And Completion
In certain cases a player could cheat and not collect all Skyshards
but update the count value, and once the last Skyshard to meet
the total was absorbed it would complete the main quest but still
leave all other objectives of unfound Skyshards in place.

This refactor checks for the total count value to be met, and when
it is it will auto-complete all objectives and mark all quests
as completed. This mirrors regular completion handling.

It also properly handles when the quest system is switched to the
lite tracking system and will auto-hide all previous objectives
and complete all relevant side and main quests.