Refactor: Overhaul Absorb Animation And Activation Handling
Previously you could activate a Skyshard multiple times which
would cause the animation effects and sound to stack, as well as
the animation being slightly off sync with the absorbing sound.

This update ensures the animations and audio only play once and
block any further attempts until the original absorption is
completed, as well as syncs up the animations to the audio better.