Refactor: Overhaul Quest Manager
With the last Quest Manager update an issue would occur where
a newly absorbed Skyshard would not register until another was
absorbed. This update addresses this as well as the logic around
it to speed up the process and skip unnecessary checks.

Once the final Skyshard in a hold side quest was found, the quest
would still keep running, and only when all Skyshards were found
would all running hold side quests be stopped. This update also
stops a hold quest once its own Skyshards have been found.

Additionally there was quite a bit of repetition happening in the
hold quests that this addresses as well.