Docs: Update README Detailing The Archival Of This Version
After many years, most have moved on to the Special/Anniversary
edition of the game, and very few people still make use of this
legacy version, so it only makes sense to archive it in order to
fully focus on the latest version. You have been good, Oldrim,
may you rest in peace.